fix: adjusting names
