+++ src/components/inputs/Term/styles.ts
